# src/main/resources/openapi/chatbot-api.yaml
openapi: 3.0.3
info:
  title: Chatbot API ENSIM
  version: 1.0.0
  description: API REST exposant les fonctionnalites du chatbot Telegram.
servers:
  - url: http://localhost:9090
security:
  - ApiToken: []
  - ApiTokenQuery: []
paths:
  /weather:
    get:
      summary: Meteo courante
      parameters:
        - $ref: '#/components/parameters/City'
      responses:
        '200':
          description: Meteo formatee comme dans le chatbot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TextResponse'
        default:
          $ref: '#/components/responses/Error'
  /forecast:
    get:
      summary: Previsions a 5 jours
      parameters:
        - $ref: '#/components/parameters/City'
      responses:
        '200':
          description: Previsions formatees comme dans le chatbot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TextResponse'
        default:
          $ref: '#/components/responses/Error'
  /joke:
    get:
      summary: Blague aleatoire
      parameters:
        - name: quality
          in: query
          required: false
          description: best = la mieux notee, worst = la moins bien notee, sinon aleatoire.
          schema:
            type: string
            enum: [best, worst]
      responses:
        '200':
          description: Blague formatee comme dans le chatbot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TextResponse'
        default:
          $ref: '#/components/responses/Error'
  /jokes:
    get:
      summary: Liste ou recherche de blagues
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Blagues
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApiJoke'
        default:
          $ref: '#/components/responses/Error'
    post:
      summary: Ajout d'une blague
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JokeRequest'
      responses:
        '201':
          description: Blague creee
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiJoke'
        default:
          $ref: '#/components/responses/Error'
  /jokes/{id}:
    get:
      summary: Detail d'une blague
      parameters:
        - $ref: '#/components/parameters/JokeId'
      responses:
        '200':
          description: Blague
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiJoke'
        '404':
          $ref: '#/components/responses/Error'
    put:
      summary: Modification d'une blague
      parameters:
        - $ref: '#/components/parameters/JokeId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JokeRequest'
      responses:
        '200':
          description: Blague modifiee
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiJoke'
        '404':
          $ref: '#/components/responses/Error'
    delete:
      summary: Suppression d'une blague
      parameters:
        - $ref: '#/components/parameters/JokeId'
      responses:
        '204':
          description: Blague supprimee
        '404':
          $ref: '#/components/responses/Error'
  /jokes/carambar:
    get:
      summary: Blague Carambar aleatoire
      description: Recupere une blague depuis l'API externe Carambar via notre API REST.
      responses:
        '200':
          description: Blague Carambar formatee
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TextResponse'
        default:
          $ref: '#/components/responses/Error'
  /message:
    post:
      summary: Envoi d'un message Telegram
      description: reply_to_message_id permet de citer un message precis dans la conversation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
      responses:
        '200':
          description: Message envoye
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SentMessageResponse'
        default:
          $ref: '#/components/responses/Error'
  /chats:
    get:
      summary: Derniers messages recus via getUpdates
      responses:
        '200':
          description: Evenements message connus du bot
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChatEvent'
        default:
          $ref: '#/components/responses/Error'
components:
  securitySchemes:
    ApiToken:
      type: apiKey
      in: header
      name: X-API-Token
      description: Token simple. Authorization Bearer est aussi accepte par l'application.
    ApiTokenQuery:
      type: apiKey
      in: query
      name: access_token
      description: Token simple utilisable dans l'URL pour les tests navigateur.
  parameters:
    City:
      name: city
      in: query
      required: false
      schema:
        type: string
        default: Paris
    JokeId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
  responses:
    Error:
      description: Erreur applicative
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    TextResponse:
      type: object
      required: [text]
      properties:
        text:
          type: string
    ApiJoke:
      type: object
      required: [id, title, text, rating]
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
          description: Titre de la blague.
        text:
          type: string
          description: Texte de la blague.
        rating:
          type: integer
          format: int32
          minimum: 0
          maximum: 10
          description: Note sur 10.
    JokeRequest:
      type: object
      required: [title, text]
      properties:
        title:
          type: string
          minLength: 1
        text:
          type: string
          minLength: 1
        rating:
          type: integer
          format: int32
          minimum: 0
          maximum: 10
          description: Note sur 10 (defaut 5 si absente).
    SendMessageRequest:
      type: object
      required: [chat_id, text]
      properties:
        chat_id:
          type: integer
          format: int64
        text:
          type: string
          minLength: 1
        reply_to_message_id:
          type: integer
          format: int32
          description: Identifiant du message Telegram a citer.
    SentMessageResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
        message_id:
          type: integer
          format: int32
        chat_id:
          type: integer
          format: int64
        text:
          type: string
    ChatEvent:
      type: object
      required: [update_id, chat_id, message_id, text]
      properties:
        update_id:
          type: integer
          format: int32
        chat_id:
          type: integer
          format: int64
        message_id:
          type: integer
          format: int32
        text:
          type: string
    ErrorResponse:
      type: object
      required: [status, error, message]
      properties:
        status:
          type: integer
          format: int32
        error:
          type: string
        message:
          type: string